add bounties and apprenticeships placeholder pages
