+++ content/apprenticeships.md
@@ -1,0 +1,8 @@
++++
+title = "Apply for an Apprenticeship"
+date = "2021-11-25"
+weight = 1
+description = "Everything you need to know about applying for an apprenticeship"
++++
+
+WIP
