+++ content/bounties.md
@@ -1,9 +1,8 @@
 +++
-title = "Bounties"
+title = "Post a Bounty"
+date = "2021-11-25"
+weight = 1
+description = "A comprehensive guide to creating a bounty"
 +++
 
-We run a security bug bounty program with [HackerOne](https://hackerone.com). See the [Tlon Corp](https://hackerone.com/tlon/) page at HackerOne for more information. 
-
-If you've found a vulnerablity, and you'd like to claim a bounty, please submit a report there. Note that our bounty program is currently focused on [Azimuth](https://github.com/urbit/azimuth) vulnerabilities.
-
-For general inquiries and disclosures, please contact our security team at [security@tlon.io](mailto:security@tlon.io).
+WIP
